+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Fixed
 - Fixed device not rebooting after `dfu restore`.
+- Fixed BlueZ disconnecting when sending a command with `pybricksdev lwp3 repl`
+  to a City hub.
 
 ## [1.0.0-alpha.13] - 2021-08-06